Move CsvExporter to its' file
